+++ typescript/infra/config/environments/mainnet3/governance/utils.ts
@@ -3,9 +3,9 @@ import { Address } from '@hyperlane-xyz/utils';
 
 import { GovernanceType } from '../../../../src/governance.js';
 
-import { awIcasV2 } from './ica/aw2.js';
+import { awIcasLegacy } from './ica/_awLegacy.js';
+import { regularIcasLegacy } from './ica/_regularLegacy.js';
 import { awIcas } from './ica/aw.js';
-import { regularIcasV2 } from './ica/regular2.js';
 import { regularIcas } from './ica/regular.js';
 import { awSafes } from './safe/aw.js';
 import { irregularSafes } from './safe/irregular.js';
@@ -28,7 +28,7 @@ export function getGovernanceTimelocks(governanceType: GovernanceType) {
     case GovernanceType.OUSDT:
       return {};
     default:
-      throw new Error(`Unknown governance type: ${governanceType}`);
+      throw new Error(`Unsupported governance type: ${governanceType}`);
   }
 }
 
@@ -43,31 +43,31 @@ export function getGovernanceSafes(governanceType: GovernanceType) {
     case GovernanceType.OUSDT:
       return ousdtSafes;
     default:
-      throw new Error(`Unknown governance type: ${governanceType}`);
+      throw new Error(`Unsupported governance type: ${governanceType}`);
   }
 }
 
 export function getLegacyGovernanceIcas(governanceType: GovernanceType) {
   switch (governanceType) {
     case GovernanceType.Regular:
-      return regularIcas;
+      return regularIcasLegacy;
     case GovernanceType.AbacusWorks:
-      return awIcas;
+      return awIcasLegacy;
     case GovernanceType.Irregular:
       return {};
     case GovernanceType.OUSDT:
       return {};
     default:
-      throw new Error(`Unknown governance type: ${governanceType}`);
+      throw new Error(`Unsupported governance type: ${governanceType}`);
   }
 }
 
 export function getGovernanceIcas(governanceType: GovernanceType) {
   switch (governanceType) {
     case GovernanceType.Regular:
-      return regularIcasV2;
+      return regularIcas;
     case GovernanceType.AbacusWorks:
-      return awIcasV2;
+      return awIcas;
     case GovernanceType.Irregular:
       return {};
     case GovernanceType.OUSDT: