+++ typescript/sdk/src/index.ts
@@ -573,6 +573,8 @@ export {
   resolveRouterMapConfig,
   RouterAddress,
   RouterConfig,
+  MissingRouterViolation,
+  MissingEnrolledRouterViolation,
   RouterViolation,
   RouterViolationType,
 } from './router/types.js';
@@ -600,6 +602,7 @@ export {
   EvmNativeTokenAdapter,
   EvmTokenAdapter,
   EvmXERC20VSAdapter,
+  EvmXERC20Adapter,
 } from './token/adapters/EvmTokenAdapter.js';
 export {
   IHypTokenAdapter,
@@ -662,6 +665,7 @@ export {
 export { TokenMetadataMap } from './token/TokenMetadataMap.js';
 export {
   EVM_TOKEN_TYPE_TO_STANDARD,
+  LOCKBOX_STANDARDS,
   MINT_LIMITED_STANDARDS,
   PROTOCOL_TO_NATIVE_STANDARD,
   TOKEN_COLLATERALIZED_STANDARDS,
@@ -714,7 +718,8 @@ export {
   WarpRouteDeployConfigMailboxRequiredSchema,
   WarpRouteDeployConfigSchema,
   WarpRouteDeployConfigSchemaErrors,
-  XERC20LimitConfig,
+  XERC20VSLimitConfig,
+  XERC20Type,
   XERC20TokenExtraBridgesLimits,
   XERC20TokenMetadata,
 } from './token/types.js';
@@ -803,3 +808,4 @@ export {
   PROPOSER_ROLE,
 } from './timelock/evm/constants.js';
 export { EvmEventLogsReader } from './rpc/evm/EvmEventLogsReader.js';
+export { getTimelockExecutableTransactionFromBatch } from './timelock/evm/utils.js';